Fix json-decode error when consuming search-results
Since parsing of the search-result expects JSON,
make sure the API-call forces the output-format to be JSON. Otherwise, regular HTML is passed back.

And without needing the client to specify an output-format in its ProductSearchQueryConfiguration.
If the client did specify an output-format, it's overridden to be json when calling the API.