Add additives, countries, and languages taxonomy APIs (#278)
Similar to #277, this adds the rest of the taxonomies that are available from the server at the moment. It doesn't include the tag types for emb_codes, states, or traces, since states is the only one actually available, and I don't think it's that useful to fetch that as a taxonomy (it's not really user facing data, I don't think).